* commit 'cd5d61a0d359919388bc24dc5f0a97a17d9ead6d':
  Add new chart.js to controller summary replacing old image graph

+++ Gateway.Web/Database/spGetTimeStats_Result.cs
@@ -16,7 +16,7 @@ namespace Gateway.Web.Database
         public Nullable<System.DateTime> Date { get; set; }
         public string Hour { get; set; }
         public string Controller { get; set; }
-        public Nullable<int> AvgTime { get; set; }
-        public Nullable<int> MaxTime { get; set; }
+        public Nullable<long> AvgTime { get; set; }
+        public Nullable<long> MaxTime { get; set; }
     }
 }
